+++ packages/client/src/components/business/Chat/index.ts
@@ -1,1 +1,1 @@
-export { ChatItem } from './ChatItem'
+export { Chat } from './Chat'
