+++ packages/client/src/components/business/ChatItem/index.ts
@@ -1,0 +1,1 @@
+export { ChatItem } from './ChatItem'
