+++ packages/client/src/components/business/ChatItem/types.ts
@@ -1,0 +1,7 @@
+type ChatItemProps = {
+  name: string
+  image: string
+  text: string
+}
+
+export type { ChatItemProps }
