+++ packages/client/src/components/business/Message/Message.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { MessageStyle } from './StyledComponents'
-import { MessageType } from './types'
+import { MessageProps } from './types'
 
-export const Message = ({ text }: MessageType) => {
+export const Message = ({ text }: MessageProps) => {
   return <MessageStyle>{text}</MessageStyle>
 }
