+++ packages/client/src/components/business/Message/types.ts
@@ -1,3 +1,5 @@
-export type MessageType = {
+type MessageProps = {
   text: string
 }
+
+export type { MessageProps }
