+++ packages/client/src/pages/Forum/index.tsx
@@ -1,0 +1,1 @@
+export { Forum } from './Forum'
