--- packages/client/src/pages/ForumPage/types.ts
@@ -1,13 +1,0 @@
-export type MessageType = {
-  text: string
-}
-export type ChatItemType = {
-  name: string
-  image: string
-  text: string
-}
-
-export type MessagesListType = {
-  arrayOfMessages: string[]
-  isEmpty: boolean
-}
