+++ packages/client/src/pages/index.ts
@@ -5,6 +5,6 @@ export { Profile } from './Profile'
 export { ProfileEdit } from './ProfileEdit'
 export { PasswordEdit } from './PasswordEdit'
 export { Settings } from './Settings'
-export { ForumPage } from './ForumPage'
+export { Forum } from './Forum'
 export { LeaderBoard } from './LeaderBoard'
 export { PageError } from './PageError'
