+++ packages/client/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './constants'
 export * from './validations'
+export { randomId } from './randomId'
